build: add npm_integration_test && angular_integration_test
* it's tricky to get out of the runfiles tree with `bazel test` as `BUILD_WORKSPACE_DIRECTORY` is not set but I employed a trick to read the `DO_NOT_BUILD_HERE` file that is one level up from `execroot` and that contains the workspace directory. This is experimental and if `bazel test //:test.debug` fails than `bazel run` is still guaranteed to work as  `BUILD_WORKSPACE_DIRECTORY` will be set in that context

* test //integration:bazel_test and //integration:bazel-schematics_test exclusively

* run "exclusive" and "manual" bazel-in-bazel integration tests in their own CI job as they take 8m+ to execute

```
//integration:bazel-schematics_test                                      PASSED in 317.2s
//integration:bazel_test                                                 PASSED in 167.8s
```

* add & pin @babel deps as newer versions of babel break //packages/localize/src/tools/test:test

@babel/core dep had to be pinned to 7.6.4 or else //packages/localize/src/tools/test:test failed. Also //packages/localize uses @babel/generator, @babel/template, @babel/traverse & @babel/types so these deps were added to package.json as they were not being hoisted anymore from @babel/core transitive.

NB: integration/hello_world__systemjs_umd test must run with systemjs 0.20.0
NB: systemjs must be at 0.18.10 for legacy saucelabs job to pass

+++ integration/run_tests.sh
@@ -48,9 +48,23 @@ trap rm_cache EXIT
 
 for testDir in ${TEST_DIRS}; do
   [[ -d "$testDir" ]] || continue
-  echo "#################################"
+
+  # Skip all integration tests that are now handled by angular_integration_test except
+  # the tests that are tracked for payload size below.
+  if ([[ ! $testDir =~ cli-hello-world ]] && [[ ! $testDir == hello_world__closure ]]); then
+    echo ""
+    echo "######################################################################"
+    echo "NOT running integration test $testDir"
+    echo "This test is now run by bazel; see targets in /integration/BUILD.bazel"
+    echo "######################################################################"
+    continue
+  fi
+
+  echo ""
+  echo "######################################################################"
   echo "Running integration test $testDir"
-  echo "#################################"
+  echo "######################################################################"
+
   (
     cd $testDir
     rm -rf dist
